Now you can have more than one Scrapbook open at the same time, this partly resolves bug #686. Fixed a wrong declaration in pageitem.h.
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@4643 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem.h
@@ -752,7 +752,7 @@ public:
 	void setLineSpacing(double newSpacing);
 
 	/** @brief Get the PageItem-wide line spacing mode */
-	double lineSpacingMode() const { return LineSpMode; }
+	int lineSpacingMode() const { return LineSpMode; }
 	/**
 	 * @brief Set a line spacing for the frame
 	 * @param newSpacing line spacing for the frame
